requirements.txt
aws-cdk-lib==2.215.0
aws-cdk.aws-lambda-python-alpha==2.215.0a0
constructs>=10.0.0,<11.0.0
aws_cdk.lambda_layer_kubectl_v31==2.0.0
